Use abstract even loop/polling service.
This allows for replacing it with one that works well with async
frameworks in Rust.

TODO:

1. Implement this also on Mac
2. Think about getting rid of `TMdnsService::register` and replace it
with `TMdnsService::register_with_poll`, this would allow for getting
rid of the new generic `Poll` type parameter and instead using a
modified `AvahiEventLoop` directly, which wraps a plain `AvahiPoll`
instead of an `AvahiSimplePoll`.