Associate Types `TxtRecord` and `EventLoop`
The layout of the crate promises extensibility, but the traits depend on
`crate::{TxtRecord,EventLoop}` which actually are particular
implementations. This makes it impossible to e.g. provide additional
implementations for the traits on Linux, which are not directly tied to
the already existing Avahi implementation.

To resolve this, use associated types.